@@ -3,20 +3,20 @@ package lambroll
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"log"
 	"os"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/pkg/errors"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
 )
 
 var directUploadThreshold = int64(50 * 1024 * 1024) // 50MB
 
 func prepareZipfile(src string, excludes []string) (*os.File, os.FileInfo, error) {
 	if fi, err := os.Stat(src); err != nil {
-		return nil, nil, errors.Wrapf(err, "src %s is not found", src)
+		return nil, nil, fmt.Errorf("src %s is not found: %w", src, err)
 	} else if fi.IsDir() {
 		zipfile, info, err := createZipArchive(src, excludes)
 		if err != nil {
@@ -33,23 +33,23 @@ func prepareZipfile(src string, excludes []string) (*os.File, os.FileInfo, error
 	return nil, nil, fmt.Errorf("src %s is not found", src)
 }
 
-func (app *App) prepareFunctionCodeForDeploy(opt DeployOption, fn *Function) error {
-	if aws.StringValue(fn.PackageType) == packageTypeImage {
+func (app *App) prepareFunctionCodeForDeploy(ctx context.Context, opt DeployOption, fn *Function) error {
+	if fn.PackageType == types.PackageTypeImage {
 		if fn.Code == nil || fn.Code.ImageUri == nil {
-			return errors.New("PackageType=Image requires Code.ImageUri in function definition")
+			return fmt.Errorf("PackageType=Image requires Code.ImageUri in function definition")
 		}
 		// deploy docker image. no need to preprare
 		log.Printf("[info] using docker image %s", *fn.Code.ImageUri)
 
 		if fn.ImageConfig == nil {
-			fn.ImageConfig = &lambda.ImageConfig{} // reset explicitly
+			fn.ImageConfig = &types.ImageConfig{} // reset explicitly
 		}
 		return nil
 	}
 
 	if opt.SkipArchive != nil && *opt.SkipArchive {
 		if fn.Code == nil || fn.Code.S3Bucket == nil || fn.Code.S3Key == nil {
-			return errors.New("--skip-archive requires Code.S3Bucket and Code.S3key elements in function definition")
+			return fmt.Errorf("--skip-archive requires Code.S3Bucket and Code.S3key elements in function definition")
 		}
 		return nil
 	}
@@ -63,9 +63,9 @@ func (app *App) prepareFunctionCodeForDeploy(opt DeployOption, fn *Function) err
 	if fn.Code != nil {
 		if bucket, key := fn.Code.S3Bucket, fn.Code.S3Key; bucket != nil && key != nil {
 			log.Printf("[info] uploading function %d bytes to s3://%s/%s", info.Size(), *bucket, *key)
-			versionID, err := app.uploadFunctionToS3(zipfile, *bucket, *key)
+			versionID, err := app.uploadFunctionToS3(ctx, zipfile, *bucket, *key)
 			if err != nil {
-				errors.Wrapf(err, "failed to upload function zip to s3://%s/%s", *bucket, *key)
+				fmt.Errorf("failed to upload function zip to s3://%s/%s: %w", *bucket, *key, err)
 			}
 			if versionID != "" {
 				log.Printf("[info] object created as version %s", versionID)
@@ -75,37 +75,35 @@ func (app *App) prepareFunctionCodeForDeploy(opt DeployOption, fn *Function) err
 				fn.Code.S3ObjectVersion = nil
 			}
 		} else {
-			return errors.New("Code.S3Bucket or Code.S3Key are not defined")
+			return fmt.Errorf("Code.S3Bucket or Code.S3Key are not defined")
 		}
 	} else {
 		// try direct upload
 		if s := info.Size(); s > directUploadThreshold {
 			return fmt.Errorf("cannot use a zip file for update function directly. Too large file %d bytes. Please define Code.S3Bucket and Code.S3Key in function.json", s)
 		}
-		b, err := ioutil.ReadAll(zipfile)
+		b, err := io.ReadAll(zipfile)
 		if err != nil {
-			return errors.Wrap(err, "failed to read zipfile content")
+			return fmt.Errorf("failed to read zipfile content: %w", err)
 		}
-		fn.Code = &lambda.FunctionCode{ZipFile: b}
+		fn.Code = &types.FunctionCode{ZipFile: b}
 	}
 	return nil
 }
 
-func (app *App) create(opt DeployOption, fn *Function) error {
-	ctx := context.Background()
-	err := app.prepareFunctionCodeForDeploy(opt, fn)
+func (app *App) create(ctx context.Context, opt DeployOption, fn *Function) error {
+	err := app.prepareFunctionCodeForDeploy(ctx, opt, fn)
 	if err != nil {
-		return errors.Wrap(err, "failed to prepare function code")
+		return fmt.Errorf("failed to prepare function code: %w", err)
 	}
 	log.Println("[info] creating function", opt.label())
-	log.Println("[debug]\n", fn.String())
 
 	version := "(created)"
 	if !*opt.DryRun {
-		fn.Publish = opt.Publish
+		fn.Publish = *opt.Publish
 		res, err := app.createFunction(ctx, fn)
 		if err != nil {
-			return errors.Wrap(err, "failed to create function")
+			return fmt.Errorf("failed to create function: %w", err)
 		}
 		if res.Version != nil {
 			version = *res.Version
@@ -115,7 +113,7 @@ func (app *App) create(opt DeployOption, fn *Function) error {
 		}
 	}
 
-	if err := app.updateTags(fn, opt); err != nil {
+	if err := app.updateTags(ctx, fn, opt); err != nil {
 		return err
 	}
 
@@ -125,23 +123,22 @@ func (app *App) create(opt DeployOption, fn *Function) error {
 
 	log.Printf("[info] creating alias set %s to version %s %s", *opt.AliasName, version, opt.label())
 	if !*opt.DryRun {
-		alias, err := app.lambda.CreateAlias(&lambda.CreateAliasInput{
+		_, err := app.lambda.CreateAlias(ctx, &lambda.CreateAliasInput{
 			FunctionName:    fn.FunctionName,
 			FunctionVersion: aws.String(version),
 			Name:            aws.String(*opt.AliasName),
 		})
 		if err != nil {
-			return errors.Wrap(err, "failed to create alias")
+			return fmt.Errorf("failed to create alias: %w", err)
 		}
 		log.Println("[info] alias created")
-		log.Printf("[debug]\n%s", alias.String())
 	}
 	return nil
 }
 
-func (app *App) createFunction(ctx context.Context, fn *lambda.CreateFunctionInput) (*lambda.FunctionConfiguration, error) {
-	if res, err := app.lambda.CreateFunctionWithContext(ctx, fn); err != nil {
-		return nil, errors.Wrap(err, "failed to create function")
+func (app *App) createFunction(ctx context.Context, fn *lambda.CreateFunctionInput) (*lambda.CreateFunctionOutput, error) {
+	if res, err := app.lambda.CreateFunction(ctx, fn); err != nil {
+		return nil, fmt.Errorf("failed to create function: %w", err)
 	} else {
 		return res, app.waitForLastUpdateStatusSuccessful(ctx, *fn.FunctionName)
 	}
