@@ -4,14 +4,14 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"io/ioutil"
+	"errors"
+	"fmt"
 	"log"
 	"os"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/pkg/errors"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
 )
 
 // DeployOption represens an option for Deploy()
@@ -42,7 +42,7 @@ type versionAlias struct {
 
 // Expand expands ExcludeFile contents to Excludes
 func expandExcludeFile(file string) ([]string, error) {
-	b, err := ioutil.ReadFile(file)
+	b, err := os.ReadFile(file)
 	if err != nil {
 		if os.IsNotExist(err) {
 			return nil, nil
@@ -68,29 +68,26 @@ func (opt *DeployOption) String() string {
 }
 
 // Deploy deployes a new lambda function code
-func (app *App) Deploy(opt DeployOption) error {
-	ctx := context.Background()
+func (app *App) Deploy(ctx context.Context, opt DeployOption) error {
 	excludes, err := expandExcludeFile(*opt.ExcludeFile)
 	if err != nil {
-		return errors.Wrap(err, "failed to parse exclude-file")
+		return fmt.Errorf("failed to parse exclude-file: %w", err)
 	}
 	opt.Excludes = append(opt.Excludes, excludes...)
 	log.Printf("[debug] %s", opt.String())
 
 	fn, err := app.loadFunction(*opt.FunctionFilePath)
 	if err != nil {
-		return errors.Wrap(err, "failed to load function")
+		return fmt.Errorf("failed to load function: %w", err)
 	}
 
 	log.Printf("[info] starting deploy function %s", *fn.FunctionName)
-	if current, err := app.lambda.GetFunction(&lambda.GetFunctionInput{
+	if current, err := app.lambda.GetFunction(ctx, &lambda.GetFunctionInput{
 		FunctionName: fn.FunctionName,
 	}); err != nil {
-		if aerr, ok := err.(awserr.Error); ok {
-			switch aerr.Code() {
-			case lambda.ErrCodeResourceNotFoundException:
-				return app.create(opt, fn)
-			}
+		var nfe *types.ResourceNotFoundException
+		if errors.As(err, &nfe) {
+			return app.create(ctx, opt, fn)
 		}
 		return err
 	} else if err := validateUpdateFunction(current.Configuration, current.Code, fn); err != nil {
@@ -98,8 +95,8 @@ func (app *App) Deploy(opt DeployOption) error {
 	}
 	fillDefaultValues(fn)
 
-	if err := app.prepareFunctionCodeForDeploy(opt, fn); err != nil {
-		return errors.Wrap(err, "failed to prepare function code for deploy")
+	if err := app.prepareFunctionCodeForDeploy(ctx, opt, fn); err != nil {
+		return fmt.Errorf("failed to prepare function code for deploy: %w", err)
 	}
 
 	log.Println("[info] updating function configuration", opt.label())
@@ -122,26 +119,28 @@ func (app *App) Deploy(opt DeployOption) error {
 		SnapStart:         fn.SnapStart,
 	}
 	if env := fn.Environment; env == nil || env.Variables == nil {
-		confIn.Environment = &lambda.Environment{
-			Variables: map[string]*string{}, // set empty variables explicitly
+		confIn.Environment = &types.Environment{
+			Variables: map[string]string{}, // set empty variables explicitly
 		}
 	} else {
 		confIn.Environment = env
 	}
 
-	log.Printf("[debug]\n%s", confIn.String())
+	log.Printf("[debug]\n%s", ToJSONString(confIn))
 
 	var newerVersion string
 	if !*opt.DryRun {
-		if _, err := app.updateFunctionConfiguration(ctx, confIn); err != nil {
-			return errors.Wrap(err, "failed to update function configuration")
+		proc := func(ctx context.Context) error {
+			return app.updateFunctionConfiguration(ctx, confIn)
+		}
+		if err := app.ensureLastUpdateStatusSuccessful(ctx, *fn.FunctionName, "updating function configuration", proc); err != nil {
+			return fmt.Errorf("failed to update function configuration: %w", err)
 		}
 	}
-	if err := app.updateTags(fn, opt); err != nil {
+	if err := app.updateTags(ctx, fn, opt); err != nil {
 		return err
 	}
 
-	log.Println("[info] updating function code", opt.label())
 	codeIn := &lambda.UpdateFunctionCodeInput{
 		Architectures:   fn.Architectures,
 		FunctionName:    fn.FunctionName,
@@ -152,14 +151,19 @@ func (app *App) Deploy(opt DeployOption) error {
 		ImageUri:        fn.Code.ImageUri,
 	}
 	if *opt.DryRun {
-		codeIn.DryRun = aws.Bool(true)
+		codeIn.DryRun = true
 	} else {
-		codeIn.Publish = opt.Publish
+		codeIn.Publish = *opt.Publish
 	}
-	log.Printf("[debug]\n%s", codeIn.String())
 
-	res, err := app.updateFunctionCode(ctx, codeIn)
-	if err != nil {
+	var res *lambda.UpdateFunctionCodeOutput
+	proc := func(ctx context.Context) error {
+		var err error
+		// set res outside of this function
+		res, err = app.updateFunctionCode(ctx, codeIn)
+		return err
+	}
+	if err := app.ensureLastUpdateStatusSuccessful(ctx, *fn.FunctionName, "updating function code", proc); err != nil {
 		return err
 	}
 	if res.Version != nil {
@@ -173,130 +177,121 @@ func (app *App) Deploy(opt DeployOption) error {
 		return nil
 	}
 	if *opt.Publish || *opt.AliasToLatest {
-		err := app.updateAliases(*fn.FunctionName, versionAlias{newerVersion, *opt.AliasName})
+		err := app.updateAliases(ctx, *fn.FunctionName, versionAlias{newerVersion, *opt.AliasName})
 		if err != nil {
 			return err
 		}
 	}
 	if *opt.KeepVersions > 0 { // Ignore zero-value.
-		return app.deleteVersions(*fn.FunctionName, *opt.KeepVersions)
+		return app.deleteVersions(ctx, *fn.FunctionName, *opt.KeepVersions)
 	}
 	return nil
 }
 
-func (app *App) updateFunctionConfiguration(ctx context.Context, in *lambda.UpdateFunctionConfigurationInput) (*lambda.FunctionConfiguration, error) {
-	if err := app.waitForLastUpdateStatusSuccessful(ctx, *in.FunctionName); err != nil {
-		return nil, err
-	}
-
+func (app *App) updateFunctionConfiguration(ctx context.Context, in *lambda.UpdateFunctionConfigurationInput) error {
 	retrier := retryPolicy.Start(ctx)
 	for retrier.Continue() {
-		res, err := app.lambda.UpdateFunctionConfigurationWithContext(ctx, in)
+		_, err := app.lambda.UpdateFunctionConfiguration(ctx, in)
 		if err != nil {
-			if aerr, ok := err.(awserr.Error); ok {
-				switch aerr.Code() {
-				case lambda.ErrCodeResourceConflictException:
-					log.Println("[debug] retrying", aerr.Message())
-					continue
-				}
+			var rce *types.ResourceConflictException
+			if errors.As(err, &rce) {
+				log.Println("[debug] retrying", rce.Error())
+				continue
 			}
-			return nil, errors.Wrap(err, "failed to update function configuration")
+			return fmt.Errorf("failed to update function configuration: %w", err)
 		}
-		log.Println("[info] updated function configuration successfully")
-		return res, nil
+		return nil
 	}
-	return nil, errors.New("failed to update function configuration (max retries reached)")
+	return fmt.Errorf("failed to update function configuration (max retries reached)")
 }
 
-func (app *App) updateFunctionCode(ctx context.Context, in *lambda.UpdateFunctionCodeInput) (*lambda.FunctionConfiguration, error) {
-	if err := app.waitForLastUpdateStatusSuccessful(ctx, *in.FunctionName); err != nil {
-		return nil, err
-	}
-
-	var res *lambda.FunctionConfiguration
+func (app *App) updateFunctionCode(ctx context.Context, in *lambda.UpdateFunctionCodeInput) (*lambda.UpdateFunctionCodeOutput, error) {
+	var res *lambda.UpdateFunctionCodeOutput
 	retrier := retryPolicy.Start(ctx)
 	for retrier.Continue() {
 		var err error
-		res, err = app.lambda.UpdateFunctionCodeWithContext(ctx, in)
+		res, err = app.lambda.UpdateFunctionCode(ctx, in)
 		if err != nil {
-			if aerr, ok := err.(awserr.Error); ok {
-				switch aerr.Code() {
-				case lambda.ErrCodeResourceConflictException:
-					log.Println("[debug] retrying", aerr.Message())
-					continue
-				}
+			var rce *types.ResourceConflictException
+			if errors.As(err, &rce) {
+				log.Println("[debug] retrying", err)
+				continue
 			}
-			return nil, errors.Wrap(err, "failed to update function code")
+			return nil, fmt.Errorf("failed to update function code: %w", err)
 		}
-		log.Println("[info] update function code request was accepted")
 		break
 	}
-
-	if !retrier.Continue() {
-		return nil, errors.New("failed to update function code (max retries reached)")
-	}
-
-	if err := app.waitForLastUpdateStatusSuccessful(ctx, *in.FunctionName); err != nil {
-		return nil, err
-	}
-	log.Println("[info] updated function code successfully")
-
 	return res, nil
+}
+
+func (app *App) ensureLastUpdateStatusSuccessful(ctx context.Context, name string, msg string, code func(ctx context.Context) error) error {
+	log.Println("[info]", msg, "...")
+	if err := app.waitForLastUpdateStatusSuccessful(ctx, name); err != nil {
+		return err
+	}
+	if err := code(ctx); err != nil {
+		return err
+	}
+	log.Println("[info]", msg, "accepted. waiting for LastUpdateStatus to be successful.")
+	if err := app.waitForLastUpdateStatusSuccessful(ctx, name); err != nil {
+		return err
+	}
+	log.Println("[info]", msg, "successfully")
+	return nil
 }
 
 func (app *App) waitForLastUpdateStatusSuccessful(ctx context.Context, name string) error {
 	retrier := retryPolicy.Start(ctx)
 	for retrier.Continue() {
-		res, err := app.lambda.GetFunction(&lambda.GetFunctionInput{
+		res, err := app.lambda.GetFunction(ctx, &lambda.GetFunctionInput{
 			FunctionName: aws.String(name),
 		})
 		if err != nil {
 			log.Println("[warn] failed to get function, retrying", err)
 			continue
 		} else {
-			state := aws.StringValue(res.Configuration.State)
-			last := aws.StringValue(res.Configuration.LastUpdateStatus)
+			state := res.Configuration.State
+			last := res.Configuration.LastUpdateStatus
 			log.Printf("[info] State:%s LastUpdateStatus:%s", state, last)
-			if last == lambda.LastUpdateStatusSuccessful {
+			if last == types.LastUpdateStatusSuccessful {
 				return nil
 			}
-			log.Printf("[info] waiting for LastUpdateStatus %s", lambda.LastUpdateStatusSuccessful)
+			log.Printf("[info] waiting for LastUpdateStatus %s", types.LastUpdateStatusSuccessful)
 		}
 	}
-	return errors.New("max retries reached")
+	return fmt.Errorf("max retries reached")
 }
 
-func (app *App) updateAliases(functionName string, vs ...versionAlias) error {
+func (app *App) updateAliases(ctx context.Context, functionName string, vs ...versionAlias) error {
 	for _, v := range vs {
 		log.Printf("[info] updating alias set %s to version %s", v.Name, v.Version)
-		alias, err := app.lambda.UpdateAlias(&lambda.UpdateAliasInput{
+		_, err := app.lambda.UpdateAlias(ctx, &lambda.UpdateAliasInput{
 			FunctionName:    aws.String(functionName),
 			FunctionVersion: aws.String(v.Version),
 			Name:            aws.String(v.Name),
 		})
 		if err != nil {
-			if aerr, ok := err.(awserr.Error); ok {
-				switch aerr.Code() {
-				case lambda.ErrCodeResourceNotFoundException:
-					log.Printf("[info] alias %s is not found. creating alias", v.Name)
-					alias, err = app.lambda.CreateAlias(&lambda.CreateAliasInput{
-						FunctionName:    aws.String(functionName),
-						FunctionVersion: aws.String(v.Version),
-						Name:            aws.String(v.Name),
-					})
+			var nfe *types.ResourceNotFoundException
+			if errors.As(err, &nfe) {
+				log.Printf("[info] alias %s is not found. creating alias", v.Name)
+				_, err := app.lambda.CreateAlias(ctx, &lambda.CreateAliasInput{
+					FunctionName:    aws.String(functionName),
+					FunctionVersion: aws.String(v.Version),
+					Name:            aws.String(v.Name),
+				})
+				if err != nil {
+					return fmt.Errorf("failed to create alias: %w", err)
 				}
-			}
-			if err != nil {
-				return errors.Wrap(err, "failed to update alias")
+			} else {
+				return fmt.Errorf("failed to update alias: %w", err)
 			}
 		}
 		log.Println("[info] alias updated")
-		log.Printf("[debug]\n%s", alias.String())
 	}
 	return nil
 }
 
-func (app *App) deleteVersions(functionName string, keepVersions int) error {
+func (app *App) deleteVersions(ctx context.Context, functionName string, keepVersions int) error {
 	if keepVersions <= 0 {
 		log.Printf("[info] specify --keep-versions")
 		return nil
@@ -307,21 +302,17 @@ func (app *App) deleteVersions(functionName string, keepVersions int) error {
 	}
 
 	// versions will be set asc order, like 1 to N
-	versions := []*lambda.FunctionConfiguration{}
-
+	versions := []types.FunctionConfiguration{}
 	for {
-		req, resp := app.lambda.ListVersionsByFunctionRequest(params)
-		if err := req.Send(); err != nil {
-			return err
+		res, err := app.lambda.ListVersionsByFunction(ctx, params)
+		if err != nil {
+			return fmt.Errorf("failed to list versions: %w", err)
 		}
-
-		versions = append(versions, resp.Versions...)
-
-		if resp.NextMarker != nil {
-			params.Marker = resp.NextMarker
+		versions = append(versions, res.Versions...)
+		if res.NextMarker != nil {
+			params.Marker = res.NextMarker
 			continue
 		}
-
 		break
 	}
 
@@ -335,12 +326,12 @@ func (app *App) deleteVersions(functionName string, keepVersions int) error {
 		}
 
 		log.Printf("[info] deleting function version: %s", *v.Version)
-		_, err := app.lambda.DeleteFunction(&lambda.DeleteFunctionInput{
+		_, err := app.lambda.DeleteFunction(ctx, &lambda.DeleteFunctionInput{
 			FunctionName: aws.String(functionName),
 			Qualifier:    v.Version,
 		})
 		if err != nil {
-			return errors.Wrap(err, "failed to delete version")
+			return fmt.Errorf("failed to delete version: %w", err)
 		}
 	}
 
