@@ -2,6 +2,7 @@ package lambroll
 
 import (
 	"bufio"
+	"context"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
@@ -9,10 +10,10 @@ import (
 	"log"
 	"os"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/lambda"
 	"github.com/mattn/go-isatty"
-	"github.com/pkg/errors"
+
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	typesv2 "github.com/aws/aws-sdk-go-v2/service/lambda/types"
 )
 
 // InvokeOption represents option for Invoke()
@@ -24,19 +25,20 @@ type InvokeOption struct {
 }
 
 // Invoke invokes function
-func (app *App) Invoke(opt InvokeOption) error {
+func (app *App) Invoke(ctx context.Context, opt InvokeOption) error {
 	fn, err := app.loadFunction(*opt.FunctionFilePath)
 	if err != nil {
-		return errors.Wrap(err, "failed to load function")
+		return fmt.Errorf("failed to load function: %w", err)
 	}
-	var invocationType, logType *string
+	var invocationType typesv2.InvocationType
+	var logType typesv2.LogType
 	if *opt.Async {
-		invocationType = aws.String("Event")
+		invocationType = typesv2.InvocationTypeEvent
 	} else {
-		invocationType = aws.String("RequestResponse")
+		invocationType = typesv2.InvocationTypeRequestResponse
 	}
 	if *opt.LogTail {
-		logType = aws.String("Tail")
+		logType = typesv2.LogTypeTail
 	}
 
 	if isatty.IsTerminal(os.Stdin.Fd()) {
@@ -54,7 +56,7 @@ PAYLOAD:
 			if err == io.EOF {
 				break
 			}
-			return errors.Wrap(err, "failed to decode payload as JSON")
+			return fmt.Errorf("failed to decode payload as JSON: %w", err)
 		}
 		b, _ := json.Marshal(payload)
 		in := &lambda.InvokeInput{
@@ -66,8 +68,8 @@ PAYLOAD:
 		if len(*opt.Qualifier) > 0 {
 			in.Qualifier = opt.Qualifier
 		}
-		log.Println("[debug] invoking function", in.String())
-		res, err := app.lambda.Invoke(in)
+		log.Println("[debug] invoking function", in)
+		res, err := app.lambda.Invoke(ctx, in)
 		if err != nil {
 			log.Println("[error] failed to invoke function", err.Error())
 			continue PAYLOAD
@@ -76,7 +78,7 @@ PAYLOAD:
 		stdout.Write([]byte("\n"))
 		stdout.Flush()
 
-		log.Printf("[info] StatusCode:%d", *res.StatusCode)
+		log.Printf("[info] StatusCode:%d", res.StatusCode)
 		if res.ExecutedVersion != nil {
 			log.Printf("[info] ExecutionVersion:%s", *res.ExecutedVersion)
 		}
