@@ -1,6 +1,7 @@
 package lambroll
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"os"
@@ -8,16 +9,16 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/endpoints"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/aws/aws-sdk-go/service/sts"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	awsconfig "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/aws/aws-sdk-go-v2/service/sts"
 	"github.com/fujiwara/tfstate-lookup/tfstate"
 	"github.com/google/go-jsonnet"
 	"github.com/hashicorp/go-envparse"
 	"github.com/kayac/go-config"
-	"github.com/pkg/errors"
 	"github.com/shogo82148/go-retry"
 )
 
@@ -31,16 +32,18 @@ var retryPolicy = retry.Policy{
 }
 
 // Function represents configuration of Lambda function
+//type Function = lambda.CreateFunctionInput
+
 type Function = lambda.CreateFunctionInput
 
 // Tags represents tags of function
-type Tags = map[string]*string
+type Tags = map[string]string
 
-func (app *App) functionArn(name string) string {
+func (app *App) functionArn(ctx context.Context, name string) string {
 	return fmt.Sprintf(
 		"arn:aws:lambda:%s:%s:function:%s",
-		*app.sess.Config.Region,
-		app.AWSAccountID(),
+		app.awsConfig.Region,
+		app.AWSAccountID(ctx),
 		name,
 	)
 }
@@ -75,55 +78,66 @@ var (
 
 // App represents lambroll application
 type App struct {
-	sess      *session.Session
-	lambda    *lambda.Lambda
 	accountID string
 	profile   string
 	loader    *config.Loader
+
+	awsConfig aws.Config
+	lambda    *lambda.Client
 
 	extStr  map[string]string
 	extCode map[string]string
 }
 
+func newAwsConfig(ctx context.Context, opt *Option) (aws.Config, error) {
+	var region string
+	if opt.Region != nil && *opt.Region != "" {
+		region = aws.ToString(opt.Region)
+	}
+	optFuncs := []func(*awsconfig.LoadOptions) error{
+		awsconfig.WithRegion(region),
+	}
+	if opt.Endpoint != nil && *opt.Endpoint != "" {
+		customResolver := aws.EndpointResolverWithOptionsFunc(func(service, region string, options ...interface{}) (aws.Endpoint, error) {
+			if service == lambda.ServiceID || service == sts.ServiceID || service == s3.ServiceID {
+				return aws.Endpoint{
+					PartitionID:   "aws",
+					URL:           *opt.Endpoint,
+					SigningRegion: region,
+				}, nil
+			}
+			// returning EndpointNotFoundError will allow the service to fallback to it's default resolution
+			return aws.Endpoint{}, &aws.EndpointNotFoundError{}
+		})
+		optFuncs = append(optFuncs, awsconfig.WithEndpointResolverWithOptions(customResolver))
+	}
+	if opt.Profile != nil && *opt.Profile != "" {
+		optFuncs = append(optFuncs, awsconfig.WithSharedConfigProfile(*opt.Profile))
+	}
+	return awsconfig.LoadDefaultConfig(ctx, optFuncs...)
+}
+
 // New creates an application
-func New(opt *Option) (*App, error) {
+func New(ctx context.Context, opt *Option) (*App, error) {
 	for _, envfile := range *opt.Envfile {
 		if err := exportEnvFile(envfile); err != nil {
 			return nil, err
 		}
 	}
 
-	awsCfg := &aws.Config{}
-	if opt.Region != nil && *opt.Region != "" {
-		awsCfg.Region = aws.String(*opt.Region)
+	v2cfg, err := newAwsConfig(ctx, opt)
+	if err != nil {
+		return nil, err
 	}
-	if opt.Endpoint != nil && *opt.Endpoint != "" {
-		customResolverFunc := func(service, region string, optFns ...func(*endpoints.Options)) (endpoints.ResolvedEndpoint, error) {
-			switch service {
-			case endpoints.S3ServiceID, endpoints.LambdaServiceID, endpoints.StsServiceID:
-				return endpoints.ResolvedEndpoint{
-					URL: *opt.Endpoint,
-				}, nil
-			default:
-				return endpoints.DefaultResolver().EndpointFor(service, region, optFns...)
-			}
-		}
-		awsCfg.EndpointResolver = endpoints.ResolverFunc(customResolverFunc)
-	}
-	sessOpt := session.Options{
-		Config:            *awsCfg,
-		SharedConfigState: session.SharedConfigEnable,
-	}
+
 	var profile string
 	if opt.Profile != nil && *opt.Profile != "" {
-		sessOpt.Profile = *opt.Profile
 		profile = *opt.Profile
 	}
-	sess := session.Must(session.NewSessionWithOptions(sessOpt))
 
 	loader := config.New()
 	if opt.TFState != nil && *opt.TFState != "" {
-		funcs, err := tfstate.FuncMap(*opt.TFState)
+		funcs, err := tfstate.FuncMap(ctx, *opt.TFState)
 		if err != nil {
 			return nil, err
 		}
@@ -133,10 +147,10 @@ func New(opt *Option) (*App, error) {
 		prefixedFuncs := make(template.FuncMap)
 		for prefix, path := range *opt.PrefixedTFState {
 			if prefix == "" {
-				return nil, errors.New("--prefixed-tfstate option cannot have empty key")
+				return nil, fmt.Errorf("--prefixed-tfstate option cannot have empty key")
 			}
 
-			funcs, err := tfstate.FuncMap(path)
+			funcs, err := tfstate.FuncMap(ctx, path)
 			if err != nil {
 				return nil, err
 			}
@@ -149,10 +163,11 @@ func New(opt *Option) (*App, error) {
 	}
 
 	app := &App{
-		sess:    sess,
-		lambda:  lambda.New(sess),
 		profile: profile,
 		loader:  loader,
+
+		awsConfig: v2cfg,
+		lambda:    lambda.NewFromConfig(v2cfg),
 	}
 	if opt.ExtStr != nil {
 		app.extStr = *opt.ExtStr
@@ -165,12 +180,12 @@ func New(opt *Option) (*App, error) {
 }
 
 // AWSAccountID returns AWS account ID in current session
-func (app *App) AWSAccountID() string {
+func (app *App) AWSAccountID(ctx context.Context) string {
 	if app.accountID != "" {
 		return app.accountID
 	}
-	svc := sts.New(app.sess)
-	r, err := svc.GetCallerIdentity(&sts.GetCallerIdentityInput{})
+	svc := sts.NewFromConfig(app.awsConfig)
+	r, err := svc.GetCallerIdentity(ctx, &sts.GetCallerIdentityInput{})
 	if err != nil {
 		log.Println("[warn] failed to get caller identity", err)
 		return ""
@@ -209,42 +224,12 @@ func (app *App) loadFunction(path string) (*Function, error) {
 	}
 	var fn Function
 	if err := unmarshalJSON(src, &fn, path); err != nil {
-		return nil, errors.Wrapf(err, "failed to load %s", path)
+		return nil, fmt.Errorf("failed to load %s: %w", path, err)
 	}
 	return &fn, nil
 }
 
-func fillDefaultValues(fn *Function) {
-	if len(fn.Architectures) == 0 {
-		fn.Architectures = []*string{aws.String("x86_64")}
-	}
-	if fn.Description == nil {
-		fn.Description = aws.String("")
-	}
-	if fn.MemorySize == nil {
-		fn.MemorySize = aws.Int64(128)
-	}
-	if fn.TracingConfig == nil {
-		fn.TracingConfig = &lambda.TracingConfig{
-			Mode: aws.String("PassThrough"),
-		}
-	}
-	if fn.EphemeralStorage == nil {
-		fn.EphemeralStorage = &lambda.EphemeralStorage{
-			Size: aws.Int64(512),
-		}
-	}
-	if fn.Timeout == nil {
-		fn.Timeout = aws.Int64(3)
-	}
-	if fn.SnapStart == nil {
-		fn.SnapStart = &lambda.SnapStart{
-			ApplyOn: aws.String("None"),
-		}
-	}
-}
-
-func newFunctionFrom(c *lambda.FunctionConfiguration, code *lambda.FunctionCodeLocation, tags Tags) *Function {
+func newFunctionFrom(c *types.FunctionConfiguration, code *types.FunctionCodeLocation, tags Tags) *Function {
 	fn := &Function{
 		Architectures:     c.Architectures,
 		Description:       c.Description,
@@ -262,36 +247,38 @@ func newFunctionFrom(c *lambda.FunctionConfiguration, code *lambda.FunctionCodeL
 	}
 
 	if e := c.Environment; e != nil {
-		fn.Environment = &lambda.Environment{
+		fn.Environment = &types.Environment{
 			Variables: e.Variables,
 		}
 	}
 	if i := c.ImageConfigResponse; i != nil {
-		fn.ImageConfig = &lambda.ImageConfig{
-			Command:          i.ImageConfig.Command,
-			EntryPoint:       i.ImageConfig.EntryPoint,
-			WorkingDirectory: i.ImageConfig.WorkingDirectory,
+		if ic := i.ImageConfig; ic != nil {
+			fn.ImageConfig = &types.ImageConfig{
+				Command:          i.ImageConfig.Command,
+				EntryPoint:       i.ImageConfig.EntryPoint,
+				WorkingDirectory: i.ImageConfig.WorkingDirectory,
+			}
 		}
 	}
 	for _, layer := range c.Layers {
-		fn.Layers = append(fn.Layers, layer.Arn)
+		fn.Layers = append(fn.Layers, *layer.Arn)
 	}
 	if t := c.TracingConfig; t != nil {
-		fn.TracingConfig = &lambda.TracingConfig{
+		fn.TracingConfig = &types.TracingConfig{
 			Mode: t.Mode,
 		}
 	}
 	if v := c.VpcConfig; v != nil && *v.VpcId != "" {
-		fn.VpcConfig = &lambda.VpcConfig{
+		fn.VpcConfig = &types.VpcConfig{
 			SubnetIds:        v.SubnetIds,
 			SecurityGroupIds: v.SecurityGroupIds,
 		}
 	}
 
-	if (code != nil && aws.StringValue(code.RepositoryType) == "ECR") || aws.StringValue(fn.PackageType) == packageTypeImage {
+	if (code != nil && aws.ToString(code.RepositoryType) == "ECR") || fn.PackageType == types.PackageTypeImage {
 		log.Printf("[debug] Image URL=%s", *code.ImageUri)
-		fn.PackageType = aws.String(packageTypeImage)
-		fn.Code = &lambda.FunctionCode{
+		fn.PackageType = types.PackageTypeImage
+		fn.Code = &types.FunctionCode{
 			ImageUri: code.ImageUri,
 		}
 	}
@@ -301,11 +288,41 @@ func newFunctionFrom(c *lambda.FunctionConfiguration, code *lambda.FunctionCodeL
 	return fn
 }
 
-func newSnapStart(s *lambda.SnapStartResponse) *lambda.SnapStart {
+func fillDefaultValues(fn *Function) {
+	if len(fn.Architectures) == 0 {
+		fn.Architectures = []types.Architecture{types.ArchitectureX8664}
+	}
+	if fn.Description == nil {
+		fn.Description = aws.String("")
+	}
+	if fn.MemorySize == nil {
+		fn.MemorySize = aws.Int32(128)
+	}
+	if fn.TracingConfig == nil {
+		fn.TracingConfig = &types.TracingConfig{
+			Mode: types.TracingModePassThrough,
+		}
+	}
+	if fn.EphemeralStorage == nil {
+		fn.EphemeralStorage = &types.EphemeralStorage{
+			Size: aws.Int32(512),
+		}
+	}
+	if fn.Timeout == nil {
+		fn.Timeout = aws.Int32(3)
+	}
+	if fn.SnapStart == nil {
+		fn.SnapStart = &types.SnapStart{
+			ApplyOn: types.SnapStartApplyOnNone,
+		}
+	}
+}
+
+func newSnapStart(s *types.SnapStartResponse) *types.SnapStart {
 	if s == nil {
 		return nil
 	}
-	return &lambda.SnapStart{
+	return &types.SnapStart{
 		ApplyOn: s.ApplyOn,
 	}
 }
@@ -331,13 +348,13 @@ func exportEnvFile(file string) error {
 	return nil
 }
 
-var errCannotUpdateImageAndZip = errors.New("cannot update function code between Image and Zip")
+var errCannotUpdateImageAndZip = fmt.Errorf("cannot update function code between Image and Zip")
 
-func validateUpdateFunction(currentConf *lambda.FunctionConfiguration, currentCode *lambda.FunctionCodeLocation, newFn *lambda.CreateFunctionInput) error {
+func validateUpdateFunction(currentConf *types.FunctionConfiguration, currentCode *types.FunctionCodeLocation, newFn *lambda.CreateFunctionInput) error {
 	newCode := newFn.Code
 
 	// new=Image
-	if newCode != nil && newCode.ImageUri != nil || aws.StringValue(newFn.PackageType) == packageTypeImage {
+	if newCode != nil && newCode.ImageUri != nil || newFn.PackageType == packageTypeImage {
 		// current=Zip
 		if currentCode == nil || currentCode.ImageUri == nil {
 			return errCannotUpdateImageAndZip
@@ -345,7 +362,7 @@ func validateUpdateFunction(currentConf *lambda.FunctionConfiguration, currentCo
 	}
 
 	// current=Image
-	if currentCode != nil && currentCode.ImageUri != nil || aws.StringValue(currentConf.PackageType) == packageTypeImage {
+	if currentCode != nil && currentCode.ImageUri != nil || currentConf.PackageType == types.PackageTypeImage {
 		// new=Zip
 		if newCode == nil || newCode.ImageUri == nil {
 			return errCannotUpdateImageAndZip
