@@ -1,14 +1,16 @@
 package lambroll
 
 import (
+	"context"
+	"errors"
+	"fmt"
 	"log"
 	"strconv"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/pkg/errors"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
 )
 
 // RollbackOption represents option for Rollback()
@@ -26,26 +28,26 @@ func (opt RollbackOption) label() string {
 }
 
 // Rollback rollbacks function
-func (app *App) Rollback(opt RollbackOption) error {
+func (app *App) Rollback(ctx context.Context, opt RollbackOption) error {
 	fn, err := app.loadFunction(*opt.FunctionFilePath)
 	if err != nil {
-		return errors.Wrap(err, "failed to load function")
+		return fmt.Errorf("failed to load function: %w", err)
 	}
 
 	log.Printf("[info] starting rollback function %s", *fn.FunctionName)
 
-	res, err := app.lambda.GetAlias(&lambda.GetAliasInput{
+	res, err := app.lambda.GetAlias(ctx, &lambda.GetAliasInput{
 		FunctionName: fn.FunctionName,
 		Name:         aws.String(CurrentAliasName),
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to get alias")
+		return fmt.Errorf("failed to get alias: %w", err)
 	}
 
 	currentVersion := *res.FunctionVersion
 	cv, err := strconv.ParseInt(currentVersion, 10, 64)
 	if err != nil {
-		return errors.Wrapf(err, "failed to pase %s as int", currentVersion)
+		return fmt.Errorf("failed to pase %s as int: %w", currentVersion, err)
 	}
 
 	var prevVersion string
@@ -53,19 +55,18 @@ VERSIONS:
 	for v := cv - 1; v > 0; v-- {
 		log.Printf("[debug] get function version %d", v)
 		vs := strconv.FormatInt(v, 10)
-		res, err := app.lambda.GetFunction(&lambda.GetFunctionInput{
+		res, err := app.lambda.GetFunction(ctx, &lambda.GetFunctionInput{
 			FunctionName: fn.FunctionName,
 			Qualifier:    aws.String(vs),
 		})
 		if err != nil {
-			if aerr, ok := err.(awserr.Error); ok {
-				switch aerr.Code() {
-				case lambda.ErrCodeResourceNotFoundException:
-					log.Printf("[debug] version %s not found", vs)
-					continue VERSIONS
-				}
+			var nfe *types.ResourceNotFoundException
+			if errors.As(err, &nfe) {
+				log.Printf("[debug] version %s not found", vs)
+				continue VERSIONS
+			} else {
+				return fmt.Errorf("failed to get function: %w", err)
 			}
-			return errors.Wrap(err, "failed to get function")
 		}
 		prevVersion = *res.Configuration.Version
 		break
@@ -78,7 +79,7 @@ VERSIONS:
 	if *opt.DryRun {
 		return nil
 	}
-	err = app.updateAliases(*fn.FunctionName, versionAlias{Version: prevVersion, Name: CurrentAliasName})
+	err = app.updateAliases(ctx, *fn.FunctionName, versionAlias{Version: prevVersion, Name: CurrentAliasName})
 	if err != nil {
 		return err
 	}
@@ -87,18 +88,18 @@ VERSIONS:
 		return nil
 	}
 
-	return app.deleteFunctionVersion(*fn.FunctionName, currentVersion)
+	return app.deleteFunctionVersion(ctx, *fn.FunctionName, currentVersion)
 }
 
-func (app *App) deleteFunctionVersion(functionName, version string) error {
+func (app *App) deleteFunctionVersion(ctx context.Context, functionName, version string) error {
 	for {
 		log.Printf("[debug] checking aliased version")
-		res, err := app.lambda.GetAlias(&lambda.GetAliasInput{
+		res, err := app.lambda.GetAlias(ctx, &lambda.GetAliasInput{
 			FunctionName: aws.String(functionName),
 			Name:         aws.String(CurrentAliasName),
 		})
 		if err != nil {
-			return errors.Wrap(err, "failed to get alias")
+			return fmt.Errorf("failed to get alias: %w", err)
 		}
 		if *res.FunctionVersion == version {
 			log.Printf("[debug] version %s still has alias %s, retrying", version, CurrentAliasName)
@@ -108,12 +109,12 @@ func (app *App) deleteFunctionVersion(functionName, version string) error {
 		break
 	}
 	log.Printf("[info] deleting function version %s", version)
-	_, err := app.lambda.DeleteFunction(&lambda.DeleteFunctionInput{
+	_, err := app.lambda.DeleteFunction(ctx, &lambda.DeleteFunctionInput{
 		FunctionName: aws.String(functionName),
 		Qualifier:    aws.String(version),
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to delete version")
+		return fmt.Errorf("failed to delete version: %w", err)
 	}
 	return nil
 }
