@@ -4,13 +4,11 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"strings"
 
 	"github.com/Songmu/prompter"
-	"github.com/aws/aws-sdk-go/private/protocol/json/jsonutil"
 )
 
 func (app *App) saveFile(path string, b []byte, mode os.FileMode) error {
@@ -20,18 +18,23 @@ func (app *App) saveFile(path string, b []byte, mode os.FileMode) error {
 			return nil
 		}
 	}
-	return ioutil.WriteFile(path, b, mode)
+	return os.WriteFile(path, b, mode)
 }
 
 func marshalJSON(s interface{}) ([]byte, error) {
-	var buf bytes.Buffer
-	b, err := jsonutil.BuildJSON(s)
+	b, err := json.Marshal(s)
 	if err != nil {
 		return nil, err
 	}
-	json.Indent(&buf, b, "", "  ")
-	buf.WriteString("\n")
-	return buf.Bytes(), nil
+	x := make(map[string]interface{})
+	if err := json.Unmarshal(b, &x); err != nil {
+		return nil, err
+	}
+	if b, err := json.MarshalIndent(omitEmptyValues(x), "", "  "); err != nil {
+		return nil, err
+	} else {
+		return append(b, '\n'), nil
+	}
 }
 
 func unmarshalJSON(src []byte, v interface{}, path string) error {
